+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+[Unreleased]
+
+- Renamed `Expr` class to `ChartExpression`.
+
 ## [2.20.15] - 2024-09-09
 
 - Removed the `laminas-json` dependency and use internal code for expressions.
@@ -178,6 +182,7 @@ as possible. See [Security Advisory: ZF2014-01](https://framework.zend.com/secur
 
 - Initial release
 
+[Unreleased]: https://github.com/laurentmuller/HighchartsBundle/compare/1.7...HEAD
 [2.20.15]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.14...2.20.15
 [2.20.14]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.13...2.20.14
 [2.20.13]: https://github.com/laurentmuller/HighchartsBundle/compare/2.20.12...2.20.13